+++ legadilo/reading/views/search_views.py
@@ -46,7 +46,9 @@ from .list_of_articles_views import UpdateArticlesForm, update_list_of_articles
 
 class SearchForm(forms.Form):
     # Main fields.
-    q = forms.CharField(required=True, min_length=SEARCHED_TEXT_MIN_LENGTH, label=_("Search query"))
+    q = forms.CharField(
+        required=False, min_length=SEARCHED_TEXT_MIN_LENGTH, label=_("Search query")
+    )
     search_type = forms.ChoiceField(
         required=False,
         choices=constants.ArticleSearchType.choices,  # type: ignore[misc]
@@ -126,7 +128,7 @@ class SearchForm(forms.Form):
     def clean_q(self):
         q = self.cleaned_data["q"]
         q = full_sanitize(q)
-        if len(q) < self.fields["q"].min_length:  # type: ignore[attr-defined]
+        if q and len(q) < self.fields["q"].min_length:  # type: ignore[attr-defined]
             raise ValidationError(
                 f"You must at least enter {SEARCHED_TEXT_MIN_LENGTH} characters",
                 code="q-too-short-after-cleaning",